Generate EPUB files of cnx.org books (#122)
* add git-epub and git-epub-validate steps

* :bug: fix most errors
OPF items need attributes on them if XHTML files contain MathML

* :bug: fix a couple problems in calculus

* add file extensions to epub resources

* update docs and add all-git-epub steps

* skips missing modules if HACK_CNX_LOOSENESS is set

* Add link to cnx.org when module not in collection

* Move <link resource="file.mp3"> to resources dir

* list resources in OPF spine

* deduplicate resource in spine

* :bug: fix epub title and language

* use bash here-document syntax

* fix links to a collection

* Start cnxbook-digital-signal-processing-a-user-s-guide

* :bug: patch git-assemble bug for the epubs.

* Merge branch 'main' into epub

* :art: tidy up some python

* :art: typo

* fix a few errors

* :bug: fix a vouple more errors

* :bug: fix ToC links without text

* :bug: fix figures with titles

* :bug: fix remaining issues

* :bug: fix explicit "undefined" string in content-id
for content that was migrated from cnxml-0.5 to 0.6

* Additional transforms for user books, full list follows

- Change target="_window" to target="_blank" on a tags
- Remove empty display attributes on math
- Remove display on q tags
- Lift div tags out of dl tags and wrap the pair in a div
- Added gif MIME type

* :bug: comment missing images

* start ignoring errors we have chosen to ignore

* support baking books in other languages

* Remove assumption that media directory exists.

* :tada: add a generic style to the ePub

* Convert divs with display="inline" into spans

* Convert iframes to links

* Add some missing mimetypes

When type is application/octet-stream, try to use the original file
extension

* Some updates to the xslt in git-assemble-epub

- Update q @display template to only remove display="inline"
- Remove @display="block" from blockquote
- Remove pgwide attribute from tables. Update style to include width:
  100$ if pgwide > 0
- Transform h7, h8, and h9 to h6 because h6 is the smallest heading
- Transform captions within code tags into spans

* Add book author(s)

* :bug: fix unbound variable

* update graph

* :art: lint scripts so variables are defined

* :tada: add an enki-in-container command

* Small correction to ignoring errors in epub validation

* Small fix to the fix for validation

* This time git-epub-validate should really work, for real.

* generate optional NCX file

* Make resource links go to the module on cnx.org

* escape colons in filenames

* :bug: astronomy andadd ePub reader validation

* :bug: fix unbound variable

* :bug: fix CI (almost)
still need to fix linter errors

* Fix epub apple (#158)

fix epub structure for Apple books

* copy resources and style to the right locations
* fix relative links for Apple books

* :shirt: lint

* :bug: fix linting

* :bug: fix all lint errors

* :bug: fix python lint errors

* :bug: fix python test

* :art: fix coverage

* :shirt: lint

* Add license to OPF file

* :skull: temporarily remove EPUB validation

* Add collection id as alternative title in OPF metadata

* Update git-epub to escape book_title in opf file

* add support for audio wave files

* :shirt: lint

* :shirt: lint

* remove unnecessary edits

+++ dockerfiles/steps/git-fetch-metadata.bash
@@ -1,3 +1,5 @@
+parse_book_dir
+
 try cp -R "$IO_FETCHED/." "$IO_FETCH_META"
 
 # Based on https://github.com/openstax/content-synchronizer/blob/e04c05fdce7e1bbba6a61a859b38982e17b74a16/resource-synchronizer/sync.sh#L19-L32
@@ -15,13 +17,15 @@ fi
 try fetch-update-meta "$IO_FETCH_META/.git" "$IO_FETCH_META/modules" "$IO_FETCH_META/collections" "$ARG_GIT_REF" "$IO_FETCH_META/canonical.json"
 try rm -rf "$IO_FETCH_META/.git"
 
-[[ -f $(dirname "$IO_RESOURCES")/resources ]] || {
-    say "Expected $(dirname "$IO_RESOURCES")/resources directory to exist"
-}
+export HACK_CNX_LOOSENESS=1
+# CNX user books do not always contain media directory
+# Missing media files will still be caught by git-validate-references
+if [[ -d "$IO_FETCH_META/media" ]]; then
+    try fetch-map-resources "$IO_FETCH_META/modules" "$IO_FETCH_META/media" "$(dirname $IO_RESOURCES)" "$IO_UNUSED_RESOURCES"
+    # Either the media is in resources or unused-resources, this folder should be empty (-d will fail otherwise)
+    try rm -d "$IO_FETCH_META/media"
+fi
 
-try fetch-map-resources "$IO_FETCH_META/modules" "$IO_FETCH_META/media" "$(dirname $IO_RESOURCES)" "$IO_UNUSED_RESOURCES"
-# Either the media is in resources or unused-resources, this folder should be empty (-d will fail otherwise)
-try rm -d "$IO_FETCH_META/media"
 
 # Copy web styles to the resources directory created by fetch-map-resources
 style_resource_root="resources/styles"
